refactor - moving demo pages to specifc ux folder
