+++ pages/ux/UX_DIR_README.md
@@ -1,0 +1,3 @@
+Demo files for UX components belong here, versus being stored in the /test/ folder where experimental code is tested
+
+Tester files and static data helpers are to be stored in this folder so the dev team may see working examples of ux components.
